Fix app to working state
This patch is the beginning of migrating to using a session store
instead of passing data through props. Using stores is more useful in
this case since each subcomponent of our app can make changes to the
state of the user's data. This patch allows us direct access to this
data.

We will have to update the rest of the components to behave properly
with the session storage.