Fix incorrect ViewCard showing after archiving
When archiving a Card that isn't at the end of a List, the ViewCard
modal persists even though the Card should be removed. This occurs
because Svelte does not know that they are not the same object due
how each card is assigned an id.

We use the Card itself as the key for the each block. The svelte.dev
tutorial says it's not preferable to use the object itself, but it works
for now.

Link: https://svelte.dev/tutorial/keyed-each-blocks