Fix CalPal logo not showing up on production build
Previously, the `src` of `img` was `CalPal_Logo.png`, which is
different from `static/CalPal_logo.png`. It seems that building the
project results in the logo not loading due to case sensitivity issues.

This patch fixes this issue, so the logo should appear both on the dev
and production builds.