Rename all CSS classes
There are two approaches to styling a Svelte component:

1. Use the inline `style` property.
2. Use the :global() selector which ignores scoped styling.

The first option would result in the styling code mixing with the layout
code, which would be difficult to read. Also, any style changes we would
like to make on multiple instances of a component would be tedious.

The second option removes one advantage of Svelte which is scoped
styling. However, if we want to keep our styling and our layout
separate, this is the approach that should be taken. As a result, we
have to manually ensure there are no CSS class name collisions by
assigning them lengthy names.

Both approaches lead to less-than-ideal code, but we'll have to settle
for the lesser evil, Approach 2.

This patch changes all CSS class names to be prefixed with the Svelte
component's name to avoid name collisions. Also, the case is changed to
kebab-case. For example, the `leftHalf` class in `List` is now
`list-left-half`.