Increment List name if one with the same name already exists
Before, it was possible to have multiple Lists with the same name.
This has changed because of how we delete lists.

Currently, when the user clicks on the Delete List button, it calls a
function which modifies the session store. This function knows the name
of the List it was called from, so it simply goes through each of the
Lists in the session and compares their names. Any List with the same
name would be deleted. However, this means that deleting one List would
delete all other Lists with the same name.

To counteract this, we would have to make sure that all List names are
unique. This patch lays the groundwork by automatically renaming
untitled Lists as they are added using a naive approach. To guarantee
that all other List names are unique, we should apply this approach to
renaming Lists as well.

Our current approach to adding new Lists without name conflicts is
simple but naive:

    1. If there is no List named "Untitled List", the new List's name
       will be "Untitled List".
    2. If there is only one "Untitled List", the next added List will
       be "Untitled List (1)".
    3. If there is a List with a name like "Untitled List (n)", the
       next added List will be "Untitled List (n + 1)".

Such an approach means its possible for the user to name any List as
"Untitled List (10000)", so the new List would be "Untitled List
(10001)". I don't know if this is the best solution for now.

While working on this patch, I realized that there of a probably better
(?) way to handle this while removing the requirement of unique List
names.

Instead of letting the Delete List button delete the List
directly, we can possibly add a new field in each List in the database
called something like `marked_for_deletion`. The Delete List button
would set that to true, then the Lists in the session store would filter
out the marked List on page update.

I don't know if that's a better solution either since we would have to
rely on the fact that the database is updated every time the state of
the app changes (is that good or bad?). However, it removes the need to
verify that all List names are unique.