Get and post new_object_id on app state update
Since `index.svelte` calls `preload()` whenever the app's state is
updated, we can use that to ask MongoDB for a new `ObjectID()`. However,
when the user first opens the app, there is no `new_object_id` in the
session store since `index.svelte`'s `preload()` is not yet called. This
means that any newly added List or Card would not be given an
`ObjectID()`. After this though, `index.svelte`'s `preload()` is called,
so `new_object_id` appears in the session store.

The easy way to do it is to just make `_layout.svelte` do the same
thing. I did it by copy-pasting, which is ugly, but it works for now.