Notify Sidebar that a Card was unarchived
Apparently the list of archived Cards in the Sidebar is not contained
inside some List. Sidebar isn't even a part of Board.

I was trying to forward the `cardunarchived` event up to the Board,
since it's the component handling 'cardarchived' too. This is wrong.

It took an embarrassingly long time to find out, but it should be fixed
now.