Let ViewCard's date and time pickers use the database date-time fields
This patch makes it so the Due Date and Time fields edit the Card's
`card.due_date_time` field in the database, and the Reminder and Time
fields edit the Card's `card.remind_date_time` field in the database.

This works by splitting up the database fields into separate date and
time strings. We then bind the ViewCard input fields to these separate
strings.

Once changes are made, the strings of each input field are merged back
into a single string, which is then stored in the database.

This solves the problem where binding multiple input fields to one
database field would remove any previous changes.