Use the Lint GitHub Action for linting and formatting on pushes
The problem that took forever to solve with this one was the fact that
Prettier doesn't have an option to recursively traverse through the
files in a directory. Instead, you'd have to pass in a glob of files
(e.g. `src/**`).

If we let `prettier_dir: src/**`, `wearerequired/lint-action@v1.9.0`
would freak out and say that Prettier was not installed. We get around
this by passing `'"./**"'` to `prettier_args`, while leaving
`prettier_dir` at `src`.

Now both ESLint and Prettier recursively checks through each file in
`src`!