Create helper `userIdFromIdToken()` to get permanent user ID
What Google gives back to us when we use their sign in button is an ID
token, which we still have to send back to Google to verify. Once it is
verified, we get back a permanent (?) user ID that we can use to
identify our own users in the database.