Remove `isSignedIn` everywhere
We did this in f5bf3cca4bbf4f68eb92ba546cce58133dbe989a, but there's no
need to now. This is because we will instead redirecting the user to
`/board` whenever they have successfully logged in. Thus, there's no
need to keep track of whether they're actually signed in or not.