Render Google Sign-In in `onMount()`
It seems that the reason the button sometimes does not appear is because
Google's `api` JavaScript gets loaded before the DOM gets constructed,
so the button itself does not exist when the script tries to attach
itself onto the button.

By rendering the button ourselves with `onMount()`, we know at that
point that the script must have loaded (?).

This should resolve #42.