Load DOM tree and Google API script before rendering the sign-in button
This patch hopefully resolves #43.

Previously we would have to choose whether to render the button when the
Google API script loads or when `onMount()` is called. This means that
in some cases, the script would load first, or the component would load
first. If one loads without the other, the sign-in button would not
render.

Now, we listen to the `load` event of the script inside of `onMount()`.
This way, the DOM tree is constructed by the time `onMount()` is called,
and `renderSignInButton()` is only called when the `load` event is
fired.

In short, we load the DOM tree, then the API script, and then lastly
render the sign-in button.

We also show a Spinner while waiting for all the loading to happen.