Change sign-in button render behavior depending on what loads first
NOW this should resolve #43.

Previously, when the script loaded before the component, we weren't able
to listen to the `load` event.

Now, we first check if some variable of the script (`gapi`) exists by
the time `onMount()` is called. If it does exist, it means that the
script loaded before the component. If it doesn't, then the component
has loaded before the script.