Fix List not displaying objects with empty strings for names
What would happen is that the List would not consider Folders or Cards
with `""` as their names. This means that when their Title is blank and
we click on them to edit, the Folder or Card would disappear from the
List.

We fix this by checking only for the null-ness or undefined-ness of
`card_name` and `folder_name`. This makes it so `""` is not considered
false, which allows it to be shown in the List.