Let Title `<textarea>` automatically resize on input
This patch also includes style and implementation improvements.

Titles are now a transparent black, which were previously gray. This
meant that archived untitled Cards would seem to show no title since
they were the same color.

We use Sveltestrap 5.1.1's Input's `inner` prop to access the actual
HTML element to call our functions on. Previously, we had to pass in
some `id` to get the element with `querySelector()`. This has been
improved with this patch.