Let the drag and drop zones handle overflow scrolling instead
Previously, the Sveltestrap CardBody component was the scrollable
element in our Lists and Folders. Because the drag and drop zone was
considered to be fully visible (the parent wasn't), trying to drag a
Card to the top or bottom of the List wouldn't scroll for us.

Now, we make it so that drag and drop area itself is responsible for
handling the overflow and scrollbar instead. This means that
`svelte-dnd-action` is now aware that the drag and drop area is not
fully visible, and should handle dragging and dropping near the edges of
the List.

This also resolves #65.