Prevent `isDndShadowItem` from being stored in the database
`isDndShadowItem` is a field given by `svelte-dnd-action` to items that
have been dragged. Sometimes, the browser might crash and be unable to
remove this field. If you refresh after something like this happens, the
browser freaks out because it doesn't know what to do with this field
since we haven't dragged anything yet.

This patch fixes that by preventing the field from being saved in the
database at all, so the user data should remain fresh.