Prevent awaiting session update on `/board`
I don't know if this causes the page to slow down when we start dragging
and dropping stuff since `preload()` is called every time something gets
updated.

By getting rid of the `await`, we might be able to let Sapper deal with
fetching instead of forcing our `preload()` to wait for the update to
finish.