Make switch view button icon change on different view
The switch button is static when switching between views. This commit
makes the button icon change as the view changes. When the view is in
board, the calendar icon is bigger than the board and when the view is
in calendar, the board view is bigger. This makes the button more
intuitive in addition to adding a tooltip.