Let `getObjectId()` generate IDs on the client side
Making requests to the backend seems to take a lot of time, and making
one just to get a new `ObjectID` is kind of inefficient.

Also, I just found out that we can actually import Node modules into the
client side by pointing at its actual path in `node_modules`. Normally,
imports from just the name of the module works fine on the server side,
but causes the client side to complain since it can't resolve the name.
I'm not sure if there are any drawbacks to doing it this way.

Maybe we can do something similar with our database access? For now
though, it seems to make the most sense letting only our backend talk
to the database.