Sync new Cards on refresh
It feels wrong to just copy paste the same logic across the SignIn and
AddCalendar components, but I don't know how to write a helper function
that is server-client-agnostic.

The logic in the layout component differs only slightly from that in the
SignIn and AddCalendar components. The difference comes from how the
components access the data in the `session`.

We can live with this for now since it works fine.