Let changes in Calendar update the session and database
There's a strange quirk with what's happening inside Calendar.

`/board` calls its `preload()` whenever any changes to the state of the
Board component are made. This is not true for the Calendar component.

This means that we cannot rely on `preload()` to update our session
whenever changes to the state in Calendar are made.  Instead, we
manually make requests to change and update our session whenever there's
a change in state. This leads to a lot of lag.

`preload()` seems to be smarter at knowing when to make the requests
since changing titles and descriptions of Cards does not feel laggy at
all.

Ideally, we should be able to let `preload()` do all the work, but this
works for now.