Make tall Lists scrollable
This regression must have appeared when we implemented drag and drop
behavior, which required we modify our DOM tree to accommodate it.

The most important part in making sure a List shrinks to the size of the
Board is at line 45 of `src/components/Board.svelte`.

We allow the `<div>` around the `<List>` to shrink if it exceeds the
height of the Board, but prevent it from growing if not. If we simply
used `flex: 1`, the `<div>` around the `<List>` would vertically stretch
all the way, which meant that we could drag the area under the List
itself.

This is a similar fix to a0cc9221ad0a6954f879c035c6514a115b9b978c and
resolves #75.