Set const variable array containing accepted MIME file types
We declare a const array that stores the strings of accepted MIME file
types. This is what we use to check whether a file can be accepted or
not since we cannot check the file extension of the file. This may also
be more robust that just checking the file extension.