Assume that calendar sources starting with "http" are iCal URLs
There are Google Calendar IDs that do not end with
"calendar.google.com", such as the account's own personal calendar. This
means that we cannot safely assume that anything ending with that string
is indeed a Google Calendar ID.

Instead, we now assume that iCal URLs start with "http". We don't really
know yet if this is a safe assumption, but this should allow users to
sync their personal Google Calendars.