Check if a URL is really an iCal URL before parsing
Previously, the iCal parser would just parse any URL as if it were an
iCal URL. This means that sometimes, legit websites would be parsed as
calendars.

We fix this by sending a request to the URL. If the response contains a
`"Content-Type": "text/calendar"`, then it is an iCal URL. From here, we
actually begin parsing the URL for the calendar data.