Add a calendar source type selection in AddCalendar
This patch makes it so that the user can choose between inputting an
iCal URL or a Google Calendar ID. By letting it be an option visible to
the user, it signals to them that it's possible to also use Google
Calendar IDs alongside iCal URLs.

The error message shown is now also based on what was inputted in the
Calendar Source field.